docs: add plugin agent/ layer plan
Documents the problem (server/ conflates host code with sandbox assets),
the new 4-layer structure (front/server/agent/shared), migration steps
for each plugin, and status tracking.
