+++ apps/boring-macro-v2/src/plugins/macro/agent/sdk/boring_macro/_cli.py
