+++ apps/boring-macro-v2/src/plugins/macro/agent/tools/macroTools.ts
