+++ apps/boring-macro-v2/src/plugins/macro/agent/transforms/builtins/diff.py
