+++ apps/boring-macro-v2/src/plugins/macro/agent/transforms/builtins/moving_average.py
