+++ apps/boring-macro-v2/src/plugins/macro/agent/workspace-template/.agents/skills/macro-deck/SKILL.md
