+++ apps/boring-macro-v2/src/plugins/macro/agent/workspace-template/deck/intro.md
